+++ Ultraspherical/MultiplicationMatrices.h
@@ -55,7 +55,7 @@ class DirichletD2ConvertMultiplicationMatrix : public FilledBandedMatrix
 public:
     DirichletD2ConvertMultiplicationMatrix(DirichletD2ConvertMultiplicationMatrix& other);
     DirichletD2ConvertMultiplicationMatrix(vector<double> a);
-    virtual FilledRow *createRow(int);
+    virtual FilledRow *createRow(unsigned long);
 };
